requirements.txt
dotenv==0.9.9
python-dotenv==1.1.0
python-telegram-bot==13.5
requests==2.32.3
urllib3==1.26.15
